refactor(S-396): extract resolve_edit_fields to field_resolve.rs; add mutation-kill tests
helpers.rs was 1,253 LOC, exceeding the 1,100 LOC spec budget (review finding I1).
Extract resolve_edit_fields (~418 LOC) and its nested search_field helper to
src/cli/issue/field_resolve.rs; re-export via helpers.rs so callers are unchanged.
helpers.rs drops to 836 LOC; field_resolve.rs is 425 LOC.

Add tests 41–43 (Gate B boundary — mutation kill): assert that --field summary=X
without --summary, --field description=X without --description, and --field
priority=X without --priority do NOT trigger Gate B. Kills 3 missed mutants
(create.rs:408/414/429 && → ||, CI kill rate was 78%, target is 90%).
